# src/main/resources/public/api-docs/swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Mucry application API",
    "description" : "This is the murcy application API\n",
    "contact" : {
      "email" : "murcyapplication@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/api"
  } ],
...
  "paths" : {
...
    "/request/editor" : {
      "get" : {
        "tags" : [ "request" ],
        "summary" : "Get current user requests",
        "description" : "Get current user request if exists",
        "responses" : {
          "200" : {
            "description" : "editorRequest",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EditorRequest"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "request" ],
        "summary" : "Update last request",
        "description" : "Logs in in the application if the credentials are true",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DescriptionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Updated"
          },
          "403" : {
            "description" : "Not authorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "request" ],
        "summary" : "Request be editor",
        "description" : "Create new Request or append new one to previous closed one",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DescriptionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "created"
          },
          "403" : {
            "description" : "Not authorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "409" : {
            "description" : "Currently, there is one request unClosed",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/request/editor/list" : {
      "get" : {
        "tags" : [ "request" ],
        "summary" : "Get users requests",
        "description" : "Get users request",
        "parameters" : [ {
          "name" : "closed",
          "in" : "query",
          "description" : "editor request is closed",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        }, {
          "name" : "approved",
          "in" : "query",
          "description" : "editor request is approved",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Editor requests",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArrayEditorRequest"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/answer/{id}" : {
      "get" : {
        "tags" : [ "answer" ],
        "summary" : "Get given ID answer",
        "description" : "Get ansert by given answer identifier",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "answer id to be obtained",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Answers",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnswerDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "answer" ],
        "summary" : "Delete given ID answer",
        "description" : "Delete answer by given answer identifier",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "answer id to be deleted",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Deleted"
          },
          "403" : {
            "description" : "Not authorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/answer/list" : {
      "get" : {
        "tags" : [ "answer" ],
        "summary" : "Get user quiz answers",
        "description" : "Get current user quiz answers",
        "responses" : {
          "200" : {
            "description" : "Answers",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnswerDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/answer/list/{id}" : {
      "get" : {
        "tags" : [ "answer" ],
        "summary" : "Get given userId user quiz answers",
        "description" : "Get user by user identifer quiz answers",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "user id to retrieve answers",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Answers",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnswerDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "IndividualAnswerDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "questionId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "timeInMillis" : {
            "type" : "integer",
            "format" : "int64"
          },
          "submitedOptions" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        }
      },
      "ArrayIndividualAnswerDto" : {
        "type" : "array",
        "items" : {
          "$ref" : "#/components/schemas/IndividualAnswerDto"
        }
      },
      "AnswerDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "quizId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "timeInMillis" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPoints" : {
            "type" : "integer",
            "format" : "int64"
          },
          "individualAnswers" : {
            "$ref" : "#/components/schemas/ArrayIndividualAnswerDto"
          }
        }
      },
...
      "Error" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "integer"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "EditorRequest" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "description" : {
            "type" : "string"
          },
          "closed" : {
            "type" : "boolean"
          },
          "approved" : {
            "type" : "boolean"
          },
          "workflow" : {
            "$ref" : "#/components/schemas/WorkflowDto"
          },
          "lastWorkflow" : {
            "$ref" : "#/components/schemas/WorkflowDto"
          }
        }
      },
      "WorkflowDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/WorkflowStatus"
          },
          "statusDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "statusBy" : {
            "type" : "string",
            "description" : "reviewer username"
          },
          "nextWorkflow" : {
            "$ref" : "#/components/schemas/WorkflowDto"
          },
          "response" : {
            "type" : "string"
          }
        }
      },
      "WorkflowStatus" : {
        "type" : "string",
        "enum" : [ "APPROVED", "PENDING", "DENIED", "DRAFT", "DRAFT_FROM_APPROVED", "EXPIRED", "INCOMPLETE", "SCHEDULED" ]
      },
      "ArrayEditorRequest" : {
        "type" : "array",
        "items" : {
          "$ref" : "#/components/schemas/EditorRequest"
        }
      },
...
      "DescriptionRequest" : {
        "required" : [ "description" ],
        "type" : "object",
        "properties" : {
          "description" : {
            "type" : "string",
            "description" : "Request description"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}